プロジェクトにAVX指定追加。64bitでワーニングが出ていたのを修正。

+++ src/core/visual/gl/ResampleImageAVX2.cpp
@@ -297,7 +297,7 @@ struct AxisParamAVX2 {
 	}
 	// ê≥ãKâª
 	void normalizeAreaAvg( float* wstart, float* dweight, tjs_uint size, bool strip ) {
-		const int count = length_.size();
+		const int count = (const int)length_.size();
 		int dwindex = 0;
 		const __m256 epsilon = M256_EPSILON;
 		for( int i = 0; i < count; i++ ) {
@@ -349,7 +349,7 @@ struct AxisParamAVX2 {
 		}
 	}
 	void normalizeAreaAvg( float* wstart, tjs_uint32* dweight, tjs_uint size,  bool strip ) {
-		const int count = length_.size();
+		const int count = (const int)length_.size();
 #ifdef _DEBUG
 		std::vector<float,aligned_allocator<float,32> > work(size);
 #else
@@ -413,13 +413,13 @@ struct AxisParamAVX2 {
 			// é¿ç€ÇÃÉTÉCÉYÇãÅÇﬂÇÈ
 			int maxsize = 0;
 			if( strip == false ) {
-				int count = length_.size();
+				int count = (int)length_.size();
 				for( int i = 0; i < count; i++ ) {
 					int len = length_[i];
 					maxsize += ((len+7)>>3)<<3;	// 8 ÇÃî{êîâª
 				}
 			} else {
-				maxsize = weight.size();
+				maxsize = (int)weight.size();
 			}
 #ifdef _DEBUG
 			weight_.resize( maxsize+7 );
